pyproject.toml
[project]
name = "pythonproject6"
version = "0.1.0"
description = ""
authors = [
    {name = "Kseniya",email = "baharevaxen@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "xlrd (>=2.0.2,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '\.git'

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.poetry.group.dev.dependencies]
pytest-cov = "^6.2.1"

